Update README
Innovation Project 12/12/2018
Get Going with GIT
•	If this was a team effort, who was on the team?
Solo Project

•	What did you do?
Learned to set up GIT for a Project

•	What did you hope to gain?
To be able to use GIT and it’s functionality in a prototype project for distributed version control.

•	How did it go? / What did you actually gain?
Went well, was able to set up the project to use GIT and its features. 

•	What are the next steps (if any)?
Keep using it and  expand more of its functionality since development will be eventually moving to GIT for revision control and will be beneficial for the company.